Made file reading async and finished reading from XML to Domain

+++ RemakeOneYearLater/Models/Album.cs
@@ -8,7 +8,7 @@ namespace RemakeOneYearLater.Models
         public string Title { get; set; }
         public string TopBillingArtist { get; set; }
         public DateTime ReleaseDate { get; set; }
-        public decimal AverageUserRating { get; set; }
+        public float AverageUserRating { get; set; }
         public short Runtime { get; set; }
         public short Price { get; set; }
         public List<Track> Tracks { get; set; }
