+++ RemakeOneYearLater/Models/Movie.cs
@@ -11,7 +11,7 @@ namespace RemakeOneYearLater.Models
         public string Title { get; set; }
         public string Director { get; set; }
         public DateTime ReleaseDate { get; set; }
-        public decimal AverageUserRating { get; set; }
+        public float AverageUserRating { get; set; }
         public short Runtime { get; set; }
         public short Price { get; set; }
